modularized AddToCart, finished AddToCart for ItemDetail page, started cart views
